requirements.txt
fastapi
uvicorn
python-dotenv
groq
openai
contentstack
chromadb